Rename Profile Model to AccountsProfile
First step of refactor: Rename Profile to AccountsProfile.
